requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.12
django-appconf==1.0.3
django-celery==3.2.2
django-filter==2.0.0
django-haystack==2.8.1
django-mdeditor==0.1.13
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-shortuuidfield==0.1.3
django-simple-captcha==0.5.9
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
future==0.17.1
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
jieba==0.39
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.1.1
Pillow==5.4.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
redis==2.10.6
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
Whoosh==2.7.4
